refactor(test): extract API layer into dojo/test/api/ [test Phase 6,7,8,9]
Move 8 Test-domain serializers into api/serializer.py (TestSerializer re-exported
from api_v2/serializers.py for ReportGenerateSerializer; rest omitted, sole consumers
were the moved viewsets), ApiTestFilter/TestImportAPIFilter into api/filters.py, and
TestsViewSet/TestTypesViewSet/TestImportViewSet into api/views.py. api/urls.py adds
add_test_urls() preserving routes tests/test_types/test_imports + basenames
test/test_type/test_imports. Viewset re-exports omitted (would cycle); dojo/urls.py +
test_rest_framework updated. Finding* serializers left in place.
Full rest_framework suite green (871 tests).
