+++ dojo/test/api/__init__.py
@@ -1,0 +1,1 @@
+path = "tests"  # noqa: RUF067
